refactor: Extracted KrTabPainterAdapter interface for tab painting.
