+++ src/main/java/krasa/editorGroups/tabs2/impl/painter/KrTabPainterAdapter.kt
@@ -1,9 +1,9 @@
-package krasa.editorGroups.tabs2.impl
+package krasa.editorGroups.tabs2.impl.painter
 
-import krasa.editorGroups.tabs2.KrTabPainter
+import krasa.editorGroups.tabs2.impl.KrTabLabel
+import krasa.editorGroups.tabs2.impl.KrTabsImpl
 import krasa.editorGroups.tabs2.impl.themes.KrTabTheme
 import java.awt.Graphics
-
 
 interface KrTabPainterAdapter {
   fun paintBackground(label: KrTabLabel, g: Graphics, tabs: KrTabsImpl)
