+++ src/main/java/krasa/editorGroups/tabs2/my/EditorGroupsJBDefaultTabPainter.java
@@ -1,12 +1,12 @@
 package krasa.editorGroups.tabs2.my;
 
 import com.intellij.openapi.diagnostic.Logger;
-import krasa.editorGroups.tabs2.impl.KrDefaultTabPainter;
+import krasa.editorGroups.tabs2.impl.painter.KrDefaultTabPainter;
 
 public class EditorGroupsJBDefaultTabPainter extends KrDefaultTabPainter {
   private static final Logger LOG = com.intellij.openapi.diagnostic.Logger.getInstance(EditorGroupsJBDefaultTabPainter.class);
 
   public EditorGroupsJBDefaultTabPainter() {
-    super(new EditorGroupsTabTheme());
+    KrDefaultTabPainter(new EditorGroupsTabTheme());
   }
 }
