+++ src/main/java/krasa/editorGroups/tabs2/my/KrJBEditorTabs.java
@@ -12,9 +12,9 @@ import krasa.editorGroups.EditorGroupManager;
 import krasa.editorGroups.EditorGroupPanel;
 import krasa.editorGroups.SwitchRequest;
 import krasa.editorGroups.tabs2.KrTabInfo;
-import krasa.editorGroups.tabs2.KrTabPainter;
 import krasa.editorGroups.tabs2.impl.KrEditorTabs;
 import krasa.editorGroups.tabs2.impl.KrTabLabel;
+import krasa.editorGroups.tabs2.impl.painter.KrTabPainter;
 import krasa.editorGroups.tabs2.impl.singleRow.KrScrollableSingleRowLayout;
 import krasa.editorGroups.tabs2.impl.singleRow.KrSingleRowLayout;
 import org.jetbrains.annotations.NotNull;
